+++ modules/java/core/src/sap/commerce/toolset/java/configurator/contentEntry/SourcesContentEntryConfigurator.kt
@@ -22,6 +22,7 @@ import com.intellij.platform.backend.workspace.WorkspaceModel
 import com.intellij.platform.workspace.jps.entities.ContentRootEntityBuilder
 import com.intellij.platform.workspace.jps.entities.ModuleEntity
 import sap.commerce.toolset.java.descriptor.SourceRootEntityDescriptor
+import sap.commerce.toolset.java.descriptor.isCustomModuleDescriptor
 import sap.commerce.toolset.project.ProjectConstants
 import sap.commerce.toolset.project.context.ProjectImportContext
 import sap.commerce.toolset.project.descriptor.ModuleDescriptor
@@ -32,8 +33,10 @@ class SourcesContentEntryConfigurator : ModuleContentEntryConfigurator {
     override val name: String
         get() = "Sources"
 
-    override fun isApplicable(importContext: ProjectImportContext, moduleDescriptor: ModuleDescriptor) = moduleDescriptor.isCustomModuleDescriptor
-        || importContext.settings.importOOTBModulesInWriteMode
+    override fun isApplicable(
+        importContext: ProjectImportContext,
+        moduleDescriptor: ModuleDescriptor
+    ) = moduleDescriptor.isCustomModuleDescriptor || importContext.settings.importOOTBModulesInWriteMode
 
     override suspend fun configure(
         importContext: ProjectImportContext,