+++ modules/java/core/src/sap/commerce/toolset/java/configurator/library/HmcModuleLibraryConfigurator.kt
@@ -21,20 +21,15 @@ package sap.commerce.toolset.java.configurator.library
 import com.intellij.platform.backend.workspace.WorkspaceModel
 import com.intellij.platform.workspace.jps.entities.ModuleEntity
 import sap.commerce.toolset.java.JavaConstants
-import sap.commerce.toolset.java.configurator.library.util.compiled
-import sap.commerce.toolset.java.configurator.library.util.configureLibrary
-import sap.commerce.toolset.java.configurator.library.util.configureTestLibrary
 import sap.commerce.toolset.java.configurator.library.util.linkProjectLibrary
-import sap.commerce.toolset.project.ProjectConstants
 import sap.commerce.toolset.project.context.ProjectImportContext
 import sap.commerce.toolset.project.descriptor.ModuleDescriptor
 import sap.commerce.toolset.project.descriptor.impl.YHmcSubModuleDescriptor
-import kotlin.io.path.Path
 
-class HmcSubModuleLibraryConfigurator : ModuleLibraryConfigurator<YHmcSubModuleDescriptor> {
+class HmcModuleLibraryConfigurator : ModuleLibraryConfigurator<YHmcSubModuleDescriptor> {
 
     override val name: String
-        get() = "Hmc Libraries"
+        get() = "Hmc Sub Module"
 
     override fun isApplicable(
         importContext: ProjectImportContext,
@@ -47,30 +42,10 @@ class HmcSubModuleLibraryConfigurator : ModuleLibraryConfigurator<YHmcSubModuleD
         moduleDescriptor: YHmcSubModuleDescriptor,
         moduleEntity: ModuleEntity
     ) {
-        configureExtensionLibrary(workspaceModel, moduleDescriptor, moduleEntity)
-        configureTestLibrary(workspaceModel, moduleDescriptor, moduleEntity)
-
         moduleEntity.linkProjectLibrary(
             workspaceModel = workspaceModel,
             libraryName = JavaConstants.ProjectLibrary.HMC,
             exported = false
-        )
-    }
-
-    private suspend fun configureExtensionLibrary(
-        workspaceModel: WorkspaceModel,
-        moduleDescriptor: YHmcSubModuleDescriptor,
-        moduleEntity: ModuleEntity
-    ) {
-        val virtualFileUrlManager = workspaceModel.getVirtualFileUrlManager()
-        val libraryRoots = moduleDescriptor.compiled(
-            virtualFileUrlManager, Path(ProjectConstants.Directory.BIN)
-        )
-
-        moduleEntity.configureLibrary(
-            workspaceModel = workspaceModel,
-            libraryName = "${moduleDescriptor.name} - ${JavaConstants.ModuleLibrary.EXTENSION}",
-            libraryRoots = libraryRoots
         )
     }
 }