+++ modules/java/core/src/sap/commerce/toolset/java/configurator/library/TestModuleLibraryConfigurator.kt
@@ -19,65 +19,55 @@
 package sap.commerce.toolset.java.configurator.library
 
 import com.intellij.platform.backend.workspace.WorkspaceModel
+import com.intellij.platform.workspace.jps.entities.DependencyScope
 import com.intellij.platform.workspace.jps.entities.ModuleEntity
 import sap.commerce.toolset.java.JavaConstants
 import sap.commerce.toolset.java.configurator.library.util.*
+import sap.commerce.toolset.java.descriptor.isNonCustomModuleDescriptor
+import sap.commerce.toolset.project.ProjectConstants
 import sap.commerce.toolset.project.context.ProjectImportContext
 import sap.commerce.toolset.project.descriptor.ModuleDescriptor
-import sap.commerce.toolset.project.descriptor.impl.YHacSubModuleDescriptor
+import sap.commerce.toolset.project.descriptor.YModuleDescriptor
+import sap.commerce.toolset.util.directoryExists
 
-/**
- * https://hybris-integration.atlassian.net/browse/IIP-355
- * HAC addons can not be compiled correctly by Intellij build because
- * "hybris/bin/platform/ext/hac/web/webroot/WEB-INF/classes" from "hac" extension is not registered
- * as a dependency for HAC addons.
- */
-class HacSubModuleLibraryConfigurator : ModuleLibraryConfigurator<YHacSubModuleDescriptor> {
+class TestModuleLibraryConfigurator : ModuleLibraryConfigurator<YModuleDescriptor> {
 
     override val name: String
-        get() = "Hac Libraries"
+        get() = "Test"
 
     override fun isApplicable(
         importContext: ProjectImportContext,
         moduleDescriptor: ModuleDescriptor
-    ) = moduleDescriptor is YHacSubModuleDescriptor
+    ) = moduleDescriptor is YModuleDescriptor
 
     override suspend fun configure(
         importContext: ProjectImportContext,
         workspaceModel: WorkspaceModel,
-        moduleDescriptor: YHacSubModuleDescriptor,
-        moduleEntity: ModuleEntity
-    ) {
-        configureExtensionLibrary(importContext, workspaceModel, moduleDescriptor, moduleEntity)
-        configureTestLibrary(workspaceModel, moduleDescriptor, moduleEntity)
-
-        moduleEntity.linkProjectLibrary(
-            workspaceModel = workspaceModel,
-            libraryName = JavaConstants.ProjectLibrary.HAC,
-            exported = false,
-        )
-    }
-
-    private suspend fun configureExtensionLibrary(
-        importContext: ProjectImportContext,
-        workspaceModel: WorkspaceModel,
-        moduleDescriptor: YHacSubModuleDescriptor,
+        moduleDescriptor: YModuleDescriptor,
         moduleEntity: ModuleEntity
     ) {
         val virtualFileUrlManager = workspaceModel.getVirtualFileUrlManager()
         val libraryRoots = buildList {
-            addAll(moduleDescriptor.serverJarFiles(virtualFileUrlManager))
-            addAll(moduleDescriptor.sources(virtualFileUrlManager))
-
             addAll(moduleDescriptor.resources(virtualFileUrlManager))
-            addAll(moduleDescriptor.lib(virtualFileUrlManager))
-            addAll(moduleDescriptor.classes(virtualFileUrlManager))
-            addAll(moduleDescriptor.docSources(importContext, virtualFileUrlManager))
+
+            if (moduleDescriptor.isNonCustomModuleDescriptor) {
+                if (importContext.settings.importOOTBModulesInReadOnlyMode) {
+                    addAll(moduleDescriptor.classes(virtualFileUrlManager))
+                    addAll(moduleDescriptor.testSources(virtualFileUrlManager))
+                    addAll(moduleDescriptor.testClasses(virtualFileUrlManager))
+                } else {
+                    if (!moduleDescriptor.moduleRootPath.resolve(ProjectConstants.Directory.TEST_SRC).directoryExists) {
+                        addAll(moduleDescriptor.classes(virtualFileUrlManager))
+                        addAll(moduleDescriptor.testClasses(virtualFileUrlManager))
+                    }
+                }
+            }
         }
 
         moduleEntity.configureLibrary(
             workspaceModel = workspaceModel,
-            libraryName = "${moduleDescriptor.name} - ${JavaConstants.ModuleLibrary.EXTENSION}",
+            libraryName = "${moduleDescriptor.name} - ${JavaConstants.ModuleLibrary.TEST}",
+            scope = DependencyScope.TEST,
             libraryRoots = libraryRoots
         )
     }